project3/requirements.txt
requests>=2.28.0
click>=8.1.0
python-dotenv>=0.19.0
tabulate>=0.9.0
colorama>=0.4.4